+++ components/frontend/src/metric/MetricTypeHeader.test.js
@@ -3,10 +3,19 @@ import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { MetricTypeHeader } from "./MetricTypeHeader";
 
-function renderMetricTypeHeader(rationale, rationaleUrls) {
+function renderMetricTypeHeader(rationale, rationaleUrls, explanation, explanationUrls) {
     render(
         <MetricTypeHeader
-            metricType={{ name: "Metric type", description: "Description", rationale: rationale, rationale_urls: rationaleUrls }}
+            metricType={
+                {
+                    name: "Metric type",
+                    description: "Description",
+                    rationale: rationale,
+                    rationale_urls: rationaleUrls,
+                    explanation: explanation,
+                    explanation_urls: explanationUrls
+                }
+            }
         />
     );
 }
@@ -22,8 +31,20 @@ it('shows the rationale', async () => {
     await waitFor(() => { expect(screen.queryAllByText(/Rationale/).length).toBe(1) });
 });
 
-it('shows the rationale urls', async () => {
+it('shows the rationale URLs', async () => {
     renderMetricTypeHeader("Rationale", ["https://rationale"]);
     await userEvent.hover(screen.queryByRole("tooltip"))
     await waitFor(() => { expect(screen.queryAllByText(/https:\/\/rationale/).length).toBe(1) });
+});
+
+it('shows the explanation', async () => {
+    renderMetricTypeHeader("", [], "Explanation.");
+    await userEvent.hover(screen.queryByRole("tooltip"))
+    await waitFor(() => { expect(screen.queryAllByText(/Explanation/).length).toBe(1) });
+});
+
+it('shows the explanation URLs', async () => {
+    renderMetricTypeHeader("", [], "Explanation.", ["https://explanation"]);
+    await userEvent.hover(screen.queryByRole("tooltip"))
+    await waitFor(() => { expect(screen.queryAllByText(/https:\/\/explanation/).length).toBe(1) });
 });